+++ src/main/java/com/byunsum/ticket_reservation/payment/dto/PaymentCancelRequest.java
@@ -3,17 +3,19 @@ package com.byunsum.ticket_reservation.payment.dto;
 import com.byunsum.ticket_reservation.payment.domain.PaymentCancelReason;
 import io.swagger.v3.oas.annotations.media.Schema;
 
+import java.math.BigDecimal;
+
 public class PaymentCancelRequest {
     @Schema(description = "취소사유")
     private PaymentCancelReason reason;
 
     @Schema(description = "부분 취소 금액(전체 취소 시 null 또는 결제 금액)")
-    private Integer cancelAmount;
+    private BigDecimal cancelAmount;
 
     public PaymentCancelRequest() {
     }
 
-    public PaymentCancelRequest(PaymentCancelReason reason, Integer cancelAmount) {
+    public PaymentCancelRequest(PaymentCancelReason reason, BigDecimal cancelAmount) {
         this.reason = reason;
         this.cancelAmount = cancelAmount;
     }
@@ -22,7 +24,7 @@ public class PaymentCancelRequest {
         return reason;
     }
 
-    public Integer getCancelAmount() {
+    public BigDecimal getCancelAmount() {
         return cancelAmount;
     }
 }
