+++ src/main/java/com/byunsum/ticket_reservation/payment/repository/PaymentRepository.java
@@ -8,6 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
@@ -30,7 +31,7 @@ public interface PaymentRepository extends JpaRepository<Payment, Long> {
 
     //COALESCE 적용: null일 경우 0 반환
     @Query("SELECT coalesce(SUM(p.amount), 0) FROM Payment p WHERE p.status = 'PAID'")
-    Long getTotalPaymentAmount();
+    BigDecimal getTotalPaymentAmount();
 
     @Query("select count(p) from Payment p where p.status = 'PAID'")
     Long getTotalPaymentCount();
